+++ Leave-management/Data/ApplicationDbContext.cs
@@ -13,7 +13,7 @@ namespace Leave_management.Data
         {
         }
         public DbSet<Employee> Employees { get; set; }
-        public DbSet<LeaveHistory> LeaveHistories { get; set; }
+        public DbSet<LeaveRequest> LeaveRequests { get; set; }
         public DbSet<LeaveType> LeaveTypes { get; set; }
         public DbSet<LeaveAllocation> LeaveAllocations { get; set; }
     }
