--- Leave-management/Repository/LeaveHistoryRepository.cs
@@ -1,56 +1,0 @@
-﻿using Leave_management.Contracts;
-using Leave_management.Data;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Leave_management.Repository
-{
-    public class LeaveHistoryRepository : ILeaveHistoryRepository
-    {
-        private readonly ApplicationDbContext _db;
-        public LeaveHistoryRepository(ApplicationDbContext db)
-        {
-            _db = db;
-        }
-        public bool Create(LeaveHistory entity)
-        {
-            _db.LeaveHistories.Add(entity);
-            return Save();
-        }
-
-        public bool Delete(LeaveHistory entity)
-        {
-            _db.LeaveHistories.Remove(entity);
-            return Save();
-        }
-
-        public ICollection<LeaveHistory> FindAll()
-        {
-            return _db.LeaveHistories.ToList();
-        }
-
-        public LeaveHistory FindById(int id)
-        {
-            return _db.LeaveHistories.FirstOrDefault(x => x.Id == id);
-        }
-
-        public bool isExists(int id)
-        {
-            var isRecordExist = _db.LeaveHistories.Any(x => x.Id == id);
-            return isRecordExist;
-        }
-
-        public bool Save()
-        {
-            return _db.SaveChanges() > 0;
-        }
-
-        public bool Update(LeaveHistory entity) 
-        {
-            _db.LeaveHistories.Update(entity);
-            return Save();
-        }
-    }
-}
